+++ src/margin/mod.rs
@@ -7,26 +7,21 @@ pub mod isolated_margin_all_symbols;
 pub mod isolated_margin_disable_account;
 pub mod isolated_margin_enable_account;
 pub mod isolated_margin_fee_data;
-pub mod isolated_margin_symbol;
 pub mod isolated_margin_tier_data;
-pub mod isolated_margin_transfer;
-pub mod isolated_margin_transfer_history;
 pub mod margin_account;
+pub mod margin_account_borrow_repay;
 pub mod margin_all_assets;
 pub mod margin_all_oco_order;
 pub mod margin_all_orders;
 pub mod margin_all_pairs;
-pub mod margin_asset;
-pub mod margin_borrow;
+pub mod margin_borrow_repay_records;
 pub mod margin_cancel_oco_order;
 pub mod margin_cancel_open_orders;
 pub mod margin_cancel_order;
-pub mod margin_dustlog;
 pub mod margin_fee_data;
 pub mod margin_force_liquidation_record;
 pub mod margin_interest_history;
 pub mod margin_interest_rate_history;
-pub mod margin_loan_record;
 pub mod margin_max_borrowable;
 pub mod margin_max_transferable;
 pub mod margin_my_trades;
@@ -37,11 +32,7 @@ pub mod margin_open_oco_order;
 pub mod margin_open_orders;
 pub mod margin_order;
 pub mod margin_order_count_usage;
-pub mod margin_pair;
 pub mod margin_price_index;
-pub mod margin_repay;
-pub mod margin_repay_record;
-pub mod margin_transfer;
 pub mod margin_transfer_history;
 pub mod toggle_bnb_burn;
 
@@ -54,26 +45,21 @@ use isolated_margin_all_symbols::IsolatedMarginAllSymbols;
 use isolated_margin_disable_account::IsolatedMarginDisableAccount;
 use isolated_margin_enable_account::IsolatedMarginEnableAccount;
 use isolated_margin_fee_data::IsolatedMarginFeeData;
-use isolated_margin_symbol::IsolatedMarginSymbol;
 use isolated_margin_tier_data::IsolatedMarginTierData;
-use isolated_margin_transfer::IsolatedMarginTransfer;
-use isolated_margin_transfer_history::IsolatedMarginTransferHistory;
 use margin_account::MarginAccount;
+use margin_account_borrow_repay::MarginAccountBorrowRepay;
 use margin_all_assets::MarginAllAssets;
 use margin_all_oco_order::MarginAllOCOOrder;
 use margin_all_orders::MarginAllOrders;
 use margin_all_pairs::MarginAllPairs;
-use margin_asset::MarginAsset;
-use margin_borrow::MarginBorrow;
+use margin_borrow_repay_records::MarginBorrowRepayRecords;
 use margin_cancel_oco_order::MarginCancelOCOOrder;
 use margin_cancel_open_orders::MarginCancelOpenOrders;
 use margin_cancel_order::MarginCancelOrder;
-use margin_dustlog::MarginDustlog;
 use margin_fee_data::MarginFeeData;
 use margin_force_liquidation_record::MarginForceLiquidationRecord;
 use margin_interest_history::MarginInterestHistory;
 use margin_interest_rate_history::MarginInterestRateHistory;
-use margin_loan_record::MarginLoanRecord;
 use margin_max_borrowable::MarginMaxBorrowable;
 use margin_max_transferable::MarginMaxTransferable;
 use margin_my_trades::MarginMyTrades;
@@ -84,44 +70,12 @@ use margin_open_oco_order::MarginOpenOCOOrder;
 use margin_open_orders::MarginOpenOrders;
 use margin_order::MarginOrder;
 use margin_order_count_usage::MarginOrderCountUsage;
-use margin_pair::MarginPair;
 use margin_price_index::MarginPriceIndex;
-use margin_repay::MarginRepay;
-use margin_repay_record::MarginRepayRecord;
-use margin_transfer::MarginTransfer;
 use margin_transfer_history::MarginTransferHistory;
 use toggle_bnb_burn::ToggleBNBBurn;
 
-pub fn margin_transfer(asset: &str, amount: Decimal, r#type: u32) -> MarginTransfer {
-    MarginTransfer::new(asset, amount, r#type)
-}
-
 pub fn margin_transfer_history() -> MarginTransferHistory {
     MarginTransferHistory::new()
-}
-
-pub fn margin_borrow(asset: &str, amount: Decimal) -> MarginBorrow {
-    MarginBorrow::new(asset, amount)
-}
-
-pub fn margin_loan_record(asset: &str) -> MarginLoanRecord {
-    MarginLoanRecord::new(asset)
-}
-
-pub fn margin_repay(asset: &str, amount: Decimal) -> MarginRepay {
-    MarginRepay::new(asset, amount)
-}
-
-pub fn margin_repay_record(asset: &str) -> MarginRepayRecord {
-    MarginRepayRecord::new(asset)
-}
-
-pub fn margin_asset(asset: &str) -> MarginAsset {
-    MarginAsset::new(asset)
-}
-
-pub fn margin_pair(symbol: &str) -> MarginPair {
-    MarginPair::new(symbol)
 }
 
 pub fn margin_all_assets() -> MarginAllAssets {
@@ -154,6 +108,20 @@ pub fn margin_interest_history() -> MarginInterestHistory {
 
 pub fn margin_force_liquidation_record() -> MarginForceLiquidationRecord {
     MarginForceLiquidationRecord::new()
+}
+
+pub fn margin_account_borrow_repay(
+    asset: &str,
+    is_isolated: &str,
+    symbol: &str,
+    amount: &str,
+    type_: &str,
+) -> MarginAccountBorrowRepay {
+    MarginAccountBorrowRepay::new(asset, is_isolated, symbol, amount, type_)
+}
+
+pub fn margin_borrow_repay_records(type_: &str) -> MarginBorrowRepayRecords {
+    MarginBorrowRepayRecords::new(type_)
 }
 
 pub fn margin_account() -> MarginAccount {
@@ -210,20 +178,6 @@ pub fn margin_max_transferable(asset: &str) -> MarginMaxTransferable {
     MarginMaxTransferable::new(asset)
 }
 
-pub fn isolated_margin_transfer_history(symbol: &str) -> IsolatedMarginTransferHistory {
-    IsolatedMarginTransferHistory::new(symbol)
-}
-
-pub fn isolated_margin_transfer(
-    asset: &str,
-    symbol: &str,
-    trans_from: &str,
-    trans_to: &str,
-    amount: Decimal,
-) -> IsolatedMarginTransfer {
-    IsolatedMarginTransfer::new(asset, symbol, trans_from, trans_to, amount)
-}
-
 pub fn isolated_margin_account() -> IsolatedMarginAccount {
     IsolatedMarginAccount::new()
 }
@@ -238,10 +192,6 @@ pub fn isolated_margin_enable_account(symbol: &str) -> IsolatedMarginEnableAccou
 
 pub fn isolated_margin_account_limit() -> IsolatedMarginAccountLimit {
     IsolatedMarginAccountLimit::new()
-}
-
-pub fn isolated_margin_symbol(symbol: &str) -> IsolatedMarginSymbol {
-    IsolatedMarginSymbol::new(symbol)
 }
 
 pub fn isolated_margin_all_symbols() -> IsolatedMarginAllSymbols {
@@ -274,8 +224,4 @@ pub fn isolated_margin_tier_data(symbol: &str) -> IsolatedMarginTierData {
 
 pub fn margin_order_count_usage() -> MarginOrderCountUsage {
     MarginOrderCountUsage::new()
-}
-
-pub fn margin_dustlog() -> MarginDustlog {
-    MarginDustlog::new()
 }